bot/requirements.txt
aiogram==3.13.1
aiohttp==3.11.11
apscheduler==3.10.4
python-dotenv==1.0.1
